+++ Formula/sec-agent.rb
@@ -1,9 +1,9 @@
 class SecAgent < Formula
   desc "macOS Enclave-Bound Session Agent for Encrypted Secrets"
   homepage "https://github.com/iafilius/sec-agent"
-  url "https://github.com/iafilius/sec-agent/releases/download/v2.11.0/sec-agent_v2.11.0_darwin_arm64.tar.gz"
-  version "2.11.0"
-  sha256 "33e022755d0427dc9dcc7af0d0b55fc05b6aa66864cc96ad62cb65d40f73ca47"
+  url "https://github.com/iafilius/sec-agent/releases/download/v2.12.0/sec-agent_v2.12.0_darwin_arm64.tar.gz"
+  version "2.12.0"
+  sha256 "43e5817c20c56fed6d3b53fdbc669d18b258f6ab38abc4931a51a3eeb675a431"
   license "GPL-3.0-or-later"
 
   depends_on :macos